Refactor type string parser into own class TypeInfoParser
