+++ stratosphere-java/src/test/java/eu/stratosphere/api/java/typeutils/TypeInfoParserTest.java
@@ -47,34 +47,34 @@ public class TypeInfoParserTest {
 	
 	@Test
 	public void testBasicTypes() {
-		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, TypeInformation.parse("Integer"));
-		Assert.assertEquals(BasicTypeInfo.DOUBLE_TYPE_INFO, TypeInformation.parse("Double"));
-		Assert.assertEquals(BasicTypeInfo.BYTE_TYPE_INFO, TypeInformation.parse("Byte"));
-		Assert.assertEquals(BasicTypeInfo.FLOAT_TYPE_INFO, TypeInformation.parse("Float"));
-		Assert.assertEquals(BasicTypeInfo.SHORT_TYPE_INFO, TypeInformation.parse("Short"));
-		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, TypeInformation.parse("Long"));
-		Assert.assertEquals(BasicTypeInfo.CHAR_TYPE_INFO, TypeInformation.parse("Character"));
-		Assert.assertEquals(BasicTypeInfo.STRING_TYPE_INFO, TypeInformation.parse("String"));
-		Assert.assertEquals(BasicTypeInfo.BOOLEAN_TYPE_INFO, TypeInformation.parse("Boolean"));
+		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, TypeInfoParser.parse("Integer"));
+		Assert.assertEquals(BasicTypeInfo.DOUBLE_TYPE_INFO, TypeInfoParser.parse("Double"));
+		Assert.assertEquals(BasicTypeInfo.BYTE_TYPE_INFO, TypeInfoParser.parse("Byte"));
+		Assert.assertEquals(BasicTypeInfo.FLOAT_TYPE_INFO, TypeInfoParser.parse("Float"));
+		Assert.assertEquals(BasicTypeInfo.SHORT_TYPE_INFO, TypeInfoParser.parse("Short"));
+		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, TypeInfoParser.parse("Long"));
+		Assert.assertEquals(BasicTypeInfo.CHAR_TYPE_INFO, TypeInfoParser.parse("Character"));
+		Assert.assertEquals(BasicTypeInfo.STRING_TYPE_INFO, TypeInfoParser.parse("String"));
+		Assert.assertEquals(BasicTypeInfo.BOOLEAN_TYPE_INFO, TypeInfoParser.parse("Boolean"));
 		
-		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, TypeInformation.parse("java.lang.Integer"));
-		Assert.assertEquals(BasicTypeInfo.DOUBLE_TYPE_INFO, TypeInformation.parse("java.lang.Double"));
-		Assert.assertEquals(BasicTypeInfo.BYTE_TYPE_INFO, TypeInformation.parse("java.lang.Byte"));
-		Assert.assertEquals(BasicTypeInfo.FLOAT_TYPE_INFO, TypeInformation.parse("java.lang.Float"));
-		Assert.assertEquals(BasicTypeInfo.SHORT_TYPE_INFO, TypeInformation.parse("java.lang.Short"));
-		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, TypeInformation.parse("java.lang.Long"));
-		Assert.assertEquals(BasicTypeInfo.CHAR_TYPE_INFO, TypeInformation.parse("java.lang.Character"));
-		Assert.assertEquals(BasicTypeInfo.STRING_TYPE_INFO, TypeInformation.parse("java.lang.String"));
-		Assert.assertEquals(BasicTypeInfo.BOOLEAN_TYPE_INFO, TypeInformation.parse("java.lang.Boolean"));
+		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, TypeInfoParser.parse("java.lang.Integer"));
+		Assert.assertEquals(BasicTypeInfo.DOUBLE_TYPE_INFO, TypeInfoParser.parse("java.lang.Double"));
+		Assert.assertEquals(BasicTypeInfo.BYTE_TYPE_INFO, TypeInfoParser.parse("java.lang.Byte"));
+		Assert.assertEquals(BasicTypeInfo.FLOAT_TYPE_INFO, TypeInfoParser.parse("java.lang.Float"));
+		Assert.assertEquals(BasicTypeInfo.SHORT_TYPE_INFO, TypeInfoParser.parse("java.lang.Short"));
+		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, TypeInfoParser.parse("java.lang.Long"));
+		Assert.assertEquals(BasicTypeInfo.CHAR_TYPE_INFO, TypeInfoParser.parse("java.lang.Character"));
+		Assert.assertEquals(BasicTypeInfo.STRING_TYPE_INFO, TypeInfoParser.parse("java.lang.String"));
+		Assert.assertEquals(BasicTypeInfo.BOOLEAN_TYPE_INFO, TypeInfoParser.parse("java.lang.Boolean"));
 		
-		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, TypeInformation.parse("int"));
-		Assert.assertEquals(BasicTypeInfo.DOUBLE_TYPE_INFO, TypeInformation.parse("double"));
-		Assert.assertEquals(BasicTypeInfo.BYTE_TYPE_INFO, TypeInformation.parse("byte"));		
-		Assert.assertEquals(BasicTypeInfo.FLOAT_TYPE_INFO, TypeInformation.parse("float"));
-		Assert.assertEquals(BasicTypeInfo.SHORT_TYPE_INFO, TypeInformation.parse("short"));
-		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, TypeInformation.parse("long"));
-		Assert.assertEquals(BasicTypeInfo.CHAR_TYPE_INFO, TypeInformation.parse("char"));
-		Assert.assertEquals(BasicTypeInfo.BOOLEAN_TYPE_INFO, TypeInformation.parse("boolean"));
+		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, TypeInfoParser.parse("int"));
+		Assert.assertEquals(BasicTypeInfo.DOUBLE_TYPE_INFO, TypeInfoParser.parse("double"));
+		Assert.assertEquals(BasicTypeInfo.BYTE_TYPE_INFO, TypeInfoParser.parse("byte"));		
+		Assert.assertEquals(BasicTypeInfo.FLOAT_TYPE_INFO, TypeInfoParser.parse("float"));
+		Assert.assertEquals(BasicTypeInfo.SHORT_TYPE_INFO, TypeInfoParser.parse("short"));
+		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, TypeInfoParser.parse("long"));
+		Assert.assertEquals(BasicTypeInfo.CHAR_TYPE_INFO, TypeInfoParser.parse("char"));
+		Assert.assertEquals(BasicTypeInfo.BOOLEAN_TYPE_INFO, TypeInfoParser.parse("boolean"));
 	}
 	
 	@Test
@@ -94,7 +94,7 @@ public class TypeInfoParserTest {
 	}
 	
 	private static void helperValueType(String str, Class<?> clazz) {
-		TypeInformation<?> ti = TypeInformation.parse(str);
+		TypeInformation<?> ti = TypeInfoParser.parse(str);
 		Assert.assertTrue(ti instanceof ValueTypeInfo);
 		ValueTypeInfo<?> vti = (ValueTypeInfo<?>) ti;
 		Assert.assertEquals(clazz, vti.getTypeClass());
@@ -102,50 +102,50 @@ public class TypeInfoParserTest {
 	
 	@Test
 	public void testBasicArrays() {
-		Assert.assertEquals(BasicArrayTypeInfo.INT_ARRAY_TYPE_INFO, TypeInformation.parse("Integer[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.DOUBLE_ARRAY_TYPE_INFO, TypeInformation.parse("Double[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.BYTE_ARRAY_TYPE_INFO, TypeInformation.parse("Byte[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.FLOAT_ARRAY_TYPE_INFO, TypeInformation.parse("Float[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.SHORT_ARRAY_TYPE_INFO, TypeInformation.parse("Short[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.LONG_ARRAY_TYPE_INFO, TypeInformation.parse("Long[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.CHAR_ARRAY_TYPE_INFO, TypeInformation.parse("Character[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.STRING_ARRAY_TYPE_INFO, TypeInformation.parse("String[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.BOOLEAN_ARRAY_TYPE_INFO, TypeInformation.parse("Boolean[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.INT_ARRAY_TYPE_INFO, TypeInfoParser.parse("Integer[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.DOUBLE_ARRAY_TYPE_INFO, TypeInfoParser.parse("Double[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.BYTE_ARRAY_TYPE_INFO, TypeInfoParser.parse("Byte[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.FLOAT_ARRAY_TYPE_INFO, TypeInfoParser.parse("Float[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.SHORT_ARRAY_TYPE_INFO, TypeInfoParser.parse("Short[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.LONG_ARRAY_TYPE_INFO, TypeInfoParser.parse("Long[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.CHAR_ARRAY_TYPE_INFO, TypeInfoParser.parse("Character[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.STRING_ARRAY_TYPE_INFO, TypeInfoParser.parse("String[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.BOOLEAN_ARRAY_TYPE_INFO, TypeInfoParser.parse("Boolean[]"));
 		
-		Assert.assertEquals(BasicArrayTypeInfo.INT_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Integer[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.DOUBLE_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Double[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.BYTE_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Byte[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.FLOAT_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Float[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.SHORT_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Short[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.LONG_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Long[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.CHAR_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Character[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.STRING_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.String[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.BOOLEAN_ARRAY_TYPE_INFO, TypeInformation.parse("java.lang.Boolean[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.INT_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Integer[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.DOUBLE_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Double[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.BYTE_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Byte[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.FLOAT_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Float[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.SHORT_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Short[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.LONG_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Long[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.CHAR_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Character[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.STRING_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.String[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.BOOLEAN_ARRAY_TYPE_INFO, TypeInfoParser.parse("java.lang.Boolean[]"));
 		
-		Assert.assertEquals(BasicArrayTypeInfo.INT_ARRAY_TYPE_INFO, TypeInformation.parse("int[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.DOUBLE_ARRAY_TYPE_INFO, TypeInformation.parse("double[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.BYTE_ARRAY_TYPE_INFO, TypeInformation.parse("byte[]"));		
-		Assert.assertEquals(BasicArrayTypeInfo.FLOAT_ARRAY_TYPE_INFO, TypeInformation.parse("float[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.SHORT_ARRAY_TYPE_INFO, TypeInformation.parse("short[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.LONG_ARRAY_TYPE_INFO, TypeInformation.parse("long[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.CHAR_ARRAY_TYPE_INFO, TypeInformation.parse("char[]"));
-		Assert.assertEquals(BasicArrayTypeInfo.BOOLEAN_ARRAY_TYPE_INFO, TypeInformation.parse("boolean[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.INT_ARRAY_TYPE_INFO, TypeInfoParser.parse("int[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.DOUBLE_ARRAY_TYPE_INFO, TypeInfoParser.parse("double[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.BYTE_ARRAY_TYPE_INFO, TypeInfoParser.parse("byte[]"));		
+		Assert.assertEquals(BasicArrayTypeInfo.FLOAT_ARRAY_TYPE_INFO, TypeInfoParser.parse("float[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.SHORT_ARRAY_TYPE_INFO, TypeInfoParser.parse("short[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.LONG_ARRAY_TYPE_INFO, TypeInfoParser.parse("long[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.CHAR_ARRAY_TYPE_INFO, TypeInfoParser.parse("char[]"));
+		Assert.assertEquals(BasicArrayTypeInfo.BOOLEAN_ARRAY_TYPE_INFO, TypeInfoParser.parse("boolean[]"));
 	}
 	
 	@Test
 	public void testTuples() {
-		TypeInformation<?> ti = TypeInformation.parse("Tuple2<Integer, Long>");
+		TypeInformation<?> ti = TypeInfoParser.parse("Tuple2<Integer, Long>");
 		Assert.assertEquals(2, ti.getArity());
 		Assert.assertEquals(BasicTypeInfo.INT_TYPE_INFO, ((TupleTypeInfo<?>)ti).getTypeAt(0));
 		Assert.assertEquals(BasicTypeInfo.LONG_TYPE_INFO, ((TupleTypeInfo<?>)ti).getTypeAt(1));
 		
-		ti = TypeInformation.parse("Tuple3<Tuple1<String>, Tuple1<Integer>, Tuple2<Long, Long>>");
+		ti = TypeInfoParser.parse("Tuple3<Tuple1<String>, Tuple1<Integer>, Tuple2<Long, Long>>");
 		Assert.assertEquals("Tuple3<Tuple1<String>, Tuple1<Integer>, Tuple2<Long, Long>>", ti.toString());
 	}
 	
 	@Test
 	public void testCustomType() {
-		TypeInformation<?> ti = TypeInformation.parse("java.lang.Class");
+		TypeInformation<?> ti = TypeInfoParser.parse("java.lang.Class");
 		Assert.assertTrue(ti instanceof GenericTypeInfo);
 		Assert.assertEquals(Class.class, ((GenericTypeInfo<?>) ti).getTypeClass());
 	}
@@ -173,51 +173,51 @@ public class TypeInfoParserTest {
 	
 	@Test
 	public void testObjectArrays() {
-		TypeInformation<?> ti = TypeInformation.parse("java.lang.Class[]");
+		TypeInformation<?> ti = TypeInfoParser.parse("java.lang.Class[]");
 
 		Assert.assertTrue(ti instanceof ObjectArrayTypeInfo<?, ?>);
 		Assert.assertEquals(Class.class, ((ObjectArrayTypeInfo<?, ?>) ti).getComponentType());
 		
-		TypeInformation<?> ti2 = TypeInformation.parse("Tuple2<Integer,Double>[]");
+		TypeInformation<?> ti2 = TypeInfoParser.parse("Tuple2<Integer,Double>[]");
 
 		Assert.assertTrue(ti2 instanceof ObjectArrayTypeInfo<?, ?>);
 		Assert.assertTrue(((ObjectArrayTypeInfo<?, ?>) ti2).getComponentInfo() instanceof TupleTypeInfo);
 		
-		TypeInformation<?> ti3 = TypeInformation.parse("Tuple2<Integer[],Double>[]");
+		TypeInformation<?> ti3 = TypeInfoParser.parse("Tuple2<Integer[],Double>[]");
 		Assert.assertEquals("ObjectArrayTypeInfo<Tuple2<BasicArrayTypeInfo<Integer>, Double>>", ti3.toString());
 	}
 	
 	@Test
 	public void testLargeMixedTuple() {
-		TypeInformation<?> ti = TypeInformation.parse("eu.stratosphere.api.java.tuple.Tuple4<Double,java.lang.Class[],StringValue,Tuple1<int>>[]");
+		TypeInformation<?> ti = TypeInfoParser.parse("eu.stratosphere.api.java.tuple.Tuple4<Double,java.lang.Class[],StringValue,Tuple1<int>>[]");
 		Assert.assertEquals("ObjectArrayTypeInfo<Tuple4<Double, ObjectArrayTypeInfo<GenericType<java.lang.Class>>, ValueType<eu.stratosphere.types.StringValue>, Tuple1<Integer>>>", ti.toString());
 	}
 	
 	@Test
 	public void testException() {
 		try {
-	    	TypeInformation.parse("THIS_CLASS_DOES_NOT_EXIST");
+	    	TypeInfoParser.parse("THIS_CLASS_DOES_NOT_EXIST");
 			Assert.fail("exception expected");
 		} catch (IllegalArgumentException e) {
 			// right
 		}
 		
 		try {
-	    	TypeInformation.parse("Tuple2<Integer>");
+	    	TypeInfoParser.parse("Tuple2<Integer>");
 			Assert.fail("exception expected");
 		} catch (IllegalArgumentException e) {
 			// right
 		}
 		
 		try {
-	    	TypeInformation.parse("Tuple3<Integer,,>");
+	    	TypeInfoParser.parse("Tuple3<Integer,,>");
 			Assert.fail("exception expected");
 		} catch (IllegalArgumentException e) {
 			// right
 		}
 		
 		try {
-	    	TypeInformation.parse("Tuple1<Integer,Double>");
+	    	TypeInfoParser.parse("Tuple1<Integer,Double>");
 			Assert.fail("exception expected");
 		} catch (IllegalArgumentException e) {
 			// right
